[5주차] 심화과제 - 장바구니 페이지 작업
