+++ packages/assignment-5/src/refactoring/pages/cart/components/CartItemList/index.tsx
@@ -1,0 +1,69 @@
+import React from 'react';
+import { CartItem } from '../../../../../types';
+
+interface CartItemListProps {
+  cartItems: CartItem[];
+  updateQuantity: (productId: string, quantity: number) => void;
+  removeFromCart: (productId: string) => void;
+}
+
+const CartItemList: React.FC<CartItemListProps> = ({ cartItems, updateQuantity, removeFromCart }) => {
+  const getAppliedDiscount = (item: CartItem) => {
+    const { discounts } = item.product;
+    const { quantity } = item;
+    let appliedDiscount = 0;
+    for (const discount of discounts) {
+      if (quantity >= discount.quantity) {
+        appliedDiscount = Math.max(appliedDiscount, discount.rate);
+      }
+    }
+    return appliedDiscount;
+  };
+
+  return (
+    <div className="space-y-2">
+      {cartItems.map((item) => {
+        const appliedDiscount = getAppliedDiscount(item);
+        return (
+          <div key={item.product.id} className="flex justify-between items-center bg-white p-3 rounded shadow">
+            <div>
+              <span className="font-semibold">{item.product.name}</span>
+              <br />
+              <span className="text-sm text-gray-600">
+                {item.product.price.toLocaleString()}원 x {item.quantity}
+                {appliedDiscount > 0 && (
+                  <span className="text-green-600 ml-1">({(appliedDiscount * 100).toFixed(0)}% 할인 적용)</span>
+                )}
+              </span>
+            </div>
+            <div>
+              <button
+                onClick={() => updateQuantity(item.product.id, item.quantity - 1)}
+                className="bg-gray-300 text-gray-800 px-2 py-1 rounded mr-1 hover:bg-gray-400"
+                disabled={item.quantity <= 1}
+              >
+                -
+              </button>
+              <span className="mx-2">{item.quantity}</span>
+              <button
+                onClick={() => updateQuantity(item.product.id, item.quantity + 1)}
+                className="bg-gray-300 text-gray-800 px-2 py-1 rounded mr-1 hover:bg-gray-400"
+                disabled={item.quantity >= item.product.stock}
+              >
+                +
+              </button>
+              <button
+                onClick={() => removeFromCart(item.product.id)}
+                className="bg-red-500 text-white px-2 py-1 rounded hover:bg-red-600 ml-2"
+              >
+                삭제
+              </button>
+            </div>
+          </div>
+        );
+      })}
+    </div>
+  );
+};
+
+export default CartItemList;
